requirements.txt
numpy>=1.24
pandas>=1.5
scikit-learn>=1.2
imbalanced-learn>=0.10
lightgbm>=4.0
matplotlib>=3.7
seaborn>=0.12
pyyaml>=6.0